-- GalleryFragmentTest
RecyclerView mounted
